refactor(tarko-agent-ui): migrate stable components to ui sdk
