+++ multimodal/tarko/agent-ui/src/standalone/share/ShareModal.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
 import { FiX, FiShare2, FiLink, FiDownload, FiCheck } from 'react-icons/fi';
 import { shareService, ShareConfig, ShareResult } from './shareService';
-import { Dialog } from '@/common/components/MuiDialog';
+import { Dialog } from '@tarko/ui';
 import { LoadingSpinner } from '@/common/components/LoadingSpinner';
 
 interface ShareModalProps {
