+++ multimodal/tarko/ui/package.json
@@ -26,7 +26,8 @@
     "monaco-editor": "0.52.0",
     "parse-diff": "0.11.1",
     "@monaco-editor/react": "4.7.0",
-    "highlight.js": "^11.9.0"
+    "highlight.js": "^11.9.0",
+    "@mui/material": "^6.1.6"
   },
   "devDependencies": {
     "@rslib/core": "0.10.0",
