+++ multimodal/tarko/ui/src/components/index.ts
@@ -1,1 +1,3 @@
 export * from './code-editor';
+export * from './mui';
+export * from './dialog';
