+++ multimodal/tarko/ui/src/components/mui/index.ts
@@ -1,0 +1,1 @@
+export * from './MuiDialog';
