+++ multimodal/tarko/ui/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './file';
 export * from './pathNormalizer';
+export * from './mui';
