+++ multimodal/tarko/ui/src/utils/mui/index.ts
@@ -1,0 +1,1 @@
+export * from './muiTheme';
